访问权限bug问题已被全部修复，本地环境测试没问题了，准备把第二版部署了

+++ BackEnd/src/main/java/com/website/backend/common/security/JwtTokenProvider.java
@@ -29,7 +29,7 @@ public class JwtTokenProvider {
 	private int jwtExpirationMs;
 	
 	@Value("${app.jwtRefreshExpirationMs}")
-	private int jwtRefreshExpirationMs;
+	private long jwtRefreshExpirationMs;
 
 	public String generateToken(Authentication authentication) {
 		UserDetails userPrincipal = (UserDetails) authentication.getPrincipal();
